Split the schedule tab's header and row into their own files
ScheduleTab.kt was 1,724 lines. The header and the item row were already
top-level composables with real signatures — the header takes a count, a
density and eleven callbacks; the row takes an item, density, selection,
note and six — so both moved without inferring anything.

Six layout constants the two now share are internal rather than private:
CARD_SHAPE, ACTION_BUTTON_SIZE, ACTION_ICON_SIZE and the section variants.

ScheduleTab 604, ScheduleItemRow 561, ScheduleHeader 456.
